+++ hardware/rpiInfo/rpiInfo.h
@@ -25,7 +25,7 @@
 char* get_ip_address(void);
 void get_sd_memory(uint32_t *MemSize, uint32_t *freesize);
 void get_cpu_memory(float *Totalram, float *freeram);
-uint8_t get_temperature(void);
+float get_temperature(void);
 uint8_t get_cpu_message(void);
 uint8_t get_hard_disk_memory(uint16_t *diskMemSize, uint16_t *useMemSize);
